+++ test/js/remote/vrpcRemoteTest.js
@@ -25,8 +25,8 @@ describe('An instance of the VrpcRemote class', () => {
     assert.ok(vrpc)
   })
   it('should return available classes and functions', async () => {
-    const availabilities = await vrpc.getAvailabilities()
-    console.log('Availabilities:', JSON.stringify(availabilities, null, 2))
+    // const availabilities = await vrpc.getAvailabilities()
+    // console.log('Availabilities:', JSON.stringify(availabilities, null, 2))
     // console.log('Domains:', await vrpc.getAvailableDomains())
     // console.log('Agents:', await vrpc.getAvailableAgents())
     // console.log('Classes:', await vrpc.getAvailableClasses('js'))